Customize for me
This is my laptop. There are many like it, but this one is mine.

My laptop is my best friend. It is my life. I must master it as I must
master my life.

My laptop, without me, is useless. Without my laptop, I am useless. I
must configure my laptop true. I must code better than my peers who are
trying to code better than me. I will...

My laptop and I know that what counts in programming is not the amount
of code we write, the cleanliness of it, nor the tests that pass. We
know that it is the impact it has on the world that counts. We will
impact...

My laptop is human, even as I, because it is my life. Thus, I will learn
it as a brother. I will learn its weaknesses, its strengths, its parts,
its accessories, its keyboard and its screen. I will keep my laptop
clean and ready, even as I am clean and ready. We will become part of
each other. We will...

Before myself, I swear this creed. My laptop and I are the innovators of
my country. We are the masters of our world. We are the providers of my
life.

So be it, until laptops are obsolete.